requirements.txt
fastapi==0.111.0
psycopg2-binary==2.9.9
pydantic-settings==2.4.0
pytest==8.3.3
